requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.2.3
face_recognition==1.3.0
numpy==1.24.3
Pillow==9.5.0
python-dotenv==1.0.0
